Citizen

+++ plugins/livedesk/__plugin__/livedesk/upgrade.py
@@ -31,6 +31,7 @@ from superdesk.user.meta.user_type import UserTypeMapped
 from superdesk.user.meta.user import UserMapped
 from uuid import uuid4
 from superdesk.post.meta.post import PostMapped
+from superdesk.source.meta.type import SourceTypeMapped
 
 # --------------------------------------------------------------------
 
@@ -344,20 +345,24 @@ def upgradeSourceSmsFix():
     assert isinstance(session, Session)
 
     try:
-        session.execute('INSERT INTO source_type (`key`) values("smsblog")')
-    except (ProgrammingError, OperationalError): pass 
+        if session.query(SourceTypeMapped.id).filter(SourceTypeMapped.Key == 'smsblog').count() == 0:
+            session.execute('INSERT INTO source_type (`key`) values("smsblog")')
+    except (Exception): pass 
     
     try:
-        session.execute('INSERT INTO source_type (`key`) values("smsfeed")')
-    except (ProgrammingError, OperationalError): pass   
+        if session.query(SourceTypeMapped.id).filter(SourceTypeMapped.Key == 'smsfeed').count() == 0:
+            session.execute('INSERT INTO source_type (`key`) values("smsfeed")')
+    except (Exception): pass   
+    
+    try:   
+        idSmsfeed = session.query(SourceTypeMapped.id).filter(SourceTypeMapped.Key == 'smsfeed').scalar()
+        idFrontlineSMS = session.query(SourceTypeMapped.id).filter(SourceTypeMapped.Key == 'FrontlineSMS').scalar()     
+        session.execute('UPDATE source SET fk_type_id =' + str(idSmsfeed) + ' WHERE id=' + str(idFrontlineSMS))
+    except (Exception): pass 
     
     try:
-        session.execute('UPDATE source SET fk_type_id =  (SELECT id FROM source_type WHERE `key`="smsfeed") \
-                          WHERE fk_type_Id = (SELECT id FROM source_type WHERE `key`="FrontlineSMS")')
-    except (ProgrammingError, OperationalError): pass 
-    
-    try:
-        session.execute('DELETE FROM source_type WHERE `key` ="FrontlineSMS"')
-    except (ProgrammingError, OperationalError): pass
+        if session.query(SourceTypeMapped.id).filter(SourceTypeMapped.Key == 'FrontlineSMS').count() > 0:
+            session.execute('DELETE FROM source_type WHERE `key` ="FrontlineSMS"')
+    except (Exception): pass
        
-  
+    session.commit()
